친구,회원 목록 사진추가

+++ frontend/front/src/hooks/useFriend.jsx
@@ -1,6 +1,6 @@
 import { useState, useCallback, useEffect } from "react";
 import api from "../utils/api";
-import { showSuccessAlert, showWarningAlert } from "../utils/alertUtils.js";
+import { showSuccessAlert, showWarningAlert, showWarningDialog  } from "../utils/alertUtils.js";
 import { useAuth } from "./useAuth";
 
 export const useFriend = () => {
@@ -92,9 +92,15 @@ export const useFriend = () => {
 
 
     const deleteFriend = async (friendId) => {
+        const confirmed = await showWarningDialog({
+            title: "정말 친구를 삭제하시겠습니까?",
+            confirmText: "삭제",
+        });
+        if (!confirmed) return false;
+
         try {
             await api.delete(`/friends/${friendId}`);
-            await fetchFriends(); // 목록 새로고침
+            await fetchFriends();
             return true;
         } catch (err) {
             console.error("친구 삭제 실패:", err);
@@ -103,6 +109,12 @@ export const useFriend = () => {
     };
 
     const deleteUser = async (userId) => {
+        const confirmed = await showWarningDialog({
+            title: "정말 회원을 삭제하시겠습니까?",
+            confirmText: "삭제",
+        });
+        if (!confirmed) return false;
+
         try {
             await api.delete(`/users/${userId}`);
             return true;
